==> code/01_reformatting_text_data.ipynb <==
import os
import json
import numpy as np
import pandas as pd

directory = './data/comm_use_subset/pdf_json/'

# A function to make a list of all the json text from the files
# the "directory" should be the path to the folder containing the json pdfs
# e.g. directory = './data/comm_use_subset/pdf_json/'

def load_files(directory):
    filenames = os.listdir(directory)
    all_files = []
    for filename in filenames:
        filename = directory + filename
        file = json.load(open(filename, 'rb'))
        all_files.append(file)
    return all_files


# A function to return the discussion section of the body into its own feature

def get_discussion_text(all_files):
    discussion = []
    for i in file['body_text']:
        if i['section'].lower() == 'discussion':
            discussion.append(i['text'])
        else:
            pass
    return ''.join(discussion)


# A function to return the body, minus discussion section, into its own feature

def get_other_text(files):
    other_text = []
    for i in file['body_text']:
        if i['section'].lower() != 'discussion':
            other_text.append(i['text'])
        else:
            pass
    return ''.join(other_text)


# A function to iterate through the files to extract features and create a dataframe

def parse_text(files):
    clean_files = []
    for file in files:
        features = [
            file['paper_id'],
            file['metadata']['title'],
            get_discussion_text(file),
            get_other_text(file)
        ]
        
        clean_files.append(features)
        
    column_names = ['paper_id', 'title', 'discussion', 'text_body']
    df = pd.DataFrame(clean_files, columns = column_names)
    return df.head()

all_files = load_files('./data/comm_use_subset/pdf_json/')

df = parse_text(all_files)

## The below functions are not needed
#### They were useful in exploring how to extract features and create a dataframe

# the below, commented function returned a dictionary from the body text, but
# is no longer needed

# # A function to return a dictionary made from all the 'body' text of the files
# def get_body(file):
#     sections = [file['body_text'][i]['section'] for i in range(len(file['body_text']))]
#     body = [file['body_text'][i]['text'] for i in range(len(file['body_text']))]
#     text_dictionary = dict(zip(sections,body))
#     return text_dictionary

# A function to return a dictionary made from the 'abstract' section of the files
# def get_abstract(file):
#     for i in file
#     sections = [file['abstract'][i]['section'] for i in range(len(file['abstract']))]
#     body = [file['abstract'][i]['text'] for i in range(len(file['abstract']))]
#     abstract_dictionary = list(zip(sections,body))
#     return abstract_dictionary

# # A function to return a dictionary of authors for each journal
# def get_authors(file):
#     names = []
#     for i in range(len('authors')):   
#         first = file['metadata']['authors'][i]['first']
#         last = file['metadata']['authors'][i]['last']
#         name = ", ".join([last, first])
#         names.append(name)
#     name_dict = {}
#     name_dict['authors'] = names
#     return name_dict

# def get_author(authors):
#     names = []
#     for i in authors:
#         name = (i['first'] + ' ' + i['last'])
#         names.append(name)
#     return names
