==> code/02_filtering_dataset.ipynb <==
## Imports

import pandas as pd
import numpy as np

## Read in data

articles = pd.read_csv('datasets/metadata.csv', low_memory=False)

articles.head(2)

## Filter by coronavirus related keywords

Sources: https://www.cdc.gov/coronavirus/types.html | https://www.ncbi.nlm.nih.gov/books/NBK92442/ | https://en.wikipedia.org/wiki/Coronaviridae

We're filtering the dataframe by searching the titles and abstracts for coronavirus related keywords to get a dataframe of articles related to coronaviruses only. 

Other common human coronaviruses besides SARS, MERS, and COVID-19:
- 229E
- NL63
- OC43
- HKU1

Diseases caused by animal coronaviruses:
- PEDV: Porcine epidemic diarrhea virus (domestic pigs)
- TGEV: transmissible gastroenteritis (domestic pigs)
- FIP: feline infectious peritonitis
- IBV: infectious bronchitis virus (avian)
- MHV: Mouse hepatitis virus
- SDAV: Sialodacryoadenitis virus (mice)

# Original shape of dataframe
articles.shape

# Filter dataframe with keywords
keywords = ['coronavirus', 'covid', 'sars', 'mers', '229E', 'NL63', 'OC43', 'HKU', 'PEDV', 'TGEV', 'FIP',
            'IBV', 'MHV', 'SDAV', 'FSC']

coronavirus = articles[(articles['title'].str.contains('|'.join(keywords), case=False)) |
                  (articles['abstract'].str.contains('|'.join(keywords), case=False))].copy()

# Shape of filtered dataframe
coronavirus.shape

## Check number of missing values

# Missing values
coronavirus.isnull().sum()

## Cast publish time as datetime variable

# Datatypes of each column
coronavirus.dtypes

# Cast publish_time as datetime 
coronavirus['publish_time'] = pd.to_datetime(coronavirus['publish_time'])

coronavirus.dtypes

## Save as csv file

coronavirus.to_csv('filtered_articles.csv', index=False)
