==> code/03_merge_filtered_data.ipynb <==
## Imports

import pandas as pd
import numpy as np

## Read in csv

text = pd.read_csv('NLT_data/combined_df.csv')
articles = pd.read_csv('filtered_articles.csv')

text.head(2)

articles.head(2)

## Check dataframes

### Text dataframe

# Number of rows and cols in the text df
print(text.shape)

# Missing values in text df
text.isnull().sum()

### Filtered articles dataframe

# Number of rows and cols in the filtered articles df
print(articles.shape)

# Missing values in filtered articles df
articles.isnull().sum()

## Merge the dataframes

Since the text dataframe has no missing paper IDs, we will first join on the paper ID column in the text dataframe and the corresponding sha column in the articles dataframe. Then, we will join on the titles of both dataframes and concat the two merged dataframes into one. 

### Join on paper ID

# Inner join on sha and paper_id
corona1 = pd.merge(articles, text, how='inner', left_on='sha', right_on='paper_id')

# Number of rows
corona1.shape

corona1.head(2)

### Join on title

# Inner join on title 
corona2 = pd.merge(articles, text, how='inner', on='title')

# Number of rows
corona2.shape

corona2.head(2)

## Check for difference in merged dataframes

Before concatenating the two merged dataframes, we're going to first compare the two and clean up each dataframe accordingly.

### Missing values and differences in columns

The first merge we made was on the `paper_id` and `sha`, but since the column names were different on the two dataframes we're left with a lot of redundant columns. Let's drop the `sha` and `title_y` columns as well as any other column we don't plan to use.

corona1.isnull().sum()

For the second merged dataframe, we'll drop again drop the `sha` column since the `paper_id` column has not missing values as well as all the columns we don't plan to use.

corona2.isnull().sum()

## Drop unused and redundant columns

We're going to clean up the dataframe a bit by dropping the redundant columns as well as the columns we do not plan to use in our analysis. Then, we'll rename some columns before finally concatenating the two.

### Clean merged dataframe 1

# Drop redundant and unused columns
corona1.drop(columns=['cord_uid', 'sha', 'doi', 'pmcid', 'pubmed_id', 'license', 'Microsoft Academic Paper ID',
                      'WHO #Covidence', 'arxiv_id', 'has_pdf_parse', 'has_pmc_xml_parse', 'full_text_file',
                      'Unnamed: 0', 'title_y'], inplace=True)

# Rename columns 
corona1.rename(columns={'title_x': 'title', 'source_x': 'source'}, inplace=True)

corona1.head(2)

### Clean merged dataframe 2

# Drop unused columns
corona2.drop(columns=['cord_uid', 'sha', 'doi', 'pmcid', 'pubmed_id', 'license', 'Microsoft Academic Paper ID',
                      'WHO #Covidence', 'arxiv_id', 'has_pdf_parse', 'has_pmc_xml_parse', 'full_text_file',
                      'Unnamed: 0'], inplace=True)

# Rename columns 
corona2.rename(columns={'source_x': 'source'}, inplace=True)

corona2.head(2)

## Concatenating merged dataframes

# Concatenation
coronavirus = pd.concat([corona1, corona2], ignore_index=True)
coronavirus.head(2)

# Total number of rows
coronavirus.shape

# Missing values - no missing title or paper ids
coronavirus.isnull().sum()

# Drop any duplicates 
coronavirus.drop_duplicates(subset='paper_id', inplace=True)
coronavirus.shape

## Save as csv file

coronavirus.to_csv('merged_filtered_articles.csv', index=False)
