==> code/03_merge_metadata.ipynb <==
## Imports

import pandas as pd
import numpy as np

## Read in data

# Read in dataframes
text = pd.read_csv('combined_df.csv', low_memory=False)
articles = pd.read_csv('metadata.csv', low_memory=False)

text.head(2)

articles.head(2)

## Check for missing values

# Number of rows
text.shape

# Check for missing values
text.isnull().sum()

# Number of rows 
articles.shape

# Check for missing values
articles.isnull().sum()

## Join dataframes on title column

# Inner join on sha and paper ID
coronavirus = pd.merge(articles, text, how='inner', left_on='sha', right_on='paper_id')

# Check number of rows
coronavirus.shape

coronavirus.head(2)

# Check for missing values
coronavirus.isnull().sum()

## Drop unused rows and columns

# Drop unused columns
coronavirus.drop(columns=['cord_uid', 'doi', 'pmcid', 'pubmed_id', 'license', 'Microsoft Academic Paper ID',
                          'WHO #Covidence', 'arxiv_id', 'has_pdf_parse', 'has_pmc_xml_parse', 'full_text_file',
                          'Unnamed: 0', 'paper_id'], inplace=True)

coronavirus.head(2)

# Look at the rows with 9 missing titles
coronavirus[coronavirus['title_x'].isnull()]

# Since those 9 rows have a lot of missing values, we decided to drop them
coronavirus.dropna(subset=['title_x'], inplace=True)

coronavirus['title_x'].isnull().sum()

# Drop second title col
coronavirus.drop(columns='title_y', inplace=True)

# Rename sha, souce_x, and title_x cols
coronavirus.rename(columns={'sha': 'paper_id',
                            'source_x': 'source',
                            'title_x': 'title'},
                   inplace=True)

coronavirus.head(2)

## Check for corrupted columns

coronavirus.dtypes

## Cast publish time as datetime variable

coronavirus['publish_time'] = pd.to_datetime(coronavirus['publish_time'])

coronavirus.dtypes

## Save dataframe as csv file

coronavirus.to_csv('merged_data.csv', index=False)
