==> code/04_word2vec_metadata.ipynb <==
## Imports

import pandas as pd
import numpy as np

import regex as re
from nltk.tokenize import RegexpTokenizer
from gensim.models.word2vec import Word2Vec

## Read in data

articles = pd.read_csv('NLT_data/merged_data.csv')
articles.head(2)

## Tokenize text data

# Function that strips text of leading and trailing punctuation
import string

def strip_punc(list_of_words):
    return [word.strip(string.punctuation) for word in list_of_words]

# Function that lowercases, tokenizes, and strips the text of any trailing punctuation

def clean_text(text):
    # Cast text as string (floats in body text)
    text = str(text)
    
    # Lowercase all words
    lower = text.lower()
    
    # Tokenizes by whitespace
    tokenizer = RegexpTokenizer('\s+', gaps=True)
    tokens = tokenizer.tokenize(lower)
    
    # Strip leading and trailing punctuation
    tokens_stripped = strip_punc(tokens)
    
    # Return tokens stripped of trailing punctuation
    return tokens_stripped

# Tokenize titles
articles['title'] = articles['title'].map(clean_text)

# Tokenize abstract
articles['abstract'] = articles['abstract'].map(clean_text)

# Tokenize body text
articles['text_body'] = articles['text_body'].map(clean_text)

articles.head(2)

## Word2Vec Model - CBOW

# Train a Word2Vec model on body text using bag of words

model_cbow = Word2Vec(articles['text_body'],
                      min_count=1,
                      workers=4)

model_cbow.wv.most_similar('coronavirus', topn=25)

# Update model with words from abstract
# https://www.machinelearningplus.com/nlp/gensim-tutorial/#15howtoupdateanexistingword2vecmodelwithnewdata

model_cbow.build_vocab(articles['abstract'], update=True)
model_cbow.train(articles['abstract'],
                 total_examples = model_cbow.corpus_count,
                 epochs = model_cbow.epochs)

model_cbow.wv.most_similar('coronavirus', topn=25)

## Word2Vec Model - SkipGram

# Train Word2Vec model on body text using SkipGram

model_sg = Word2Vec(articles['text_body'],
                    min_count=1,
                    sg=1,
                    workers=4)

model_sg.wv.most_similar('coronavirus', topn=25)

# Update model with words from abstract

model_sg.build_vocab(articles['abstract'], update=True)
model_sg.train(articles['abstract'],
               total_examples = model_sg.corpus_count,
               epochs = model_sg.epochs)

model_sg.wv.most_similar('coronavirus', topn=25)
