==> code/08_add_topics.ipynb <==
## Imports

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import gensim
import gensim.corpora as corpora
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel, HdpModel
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.test.utils import common_texts
from gensim.parsing.preprocessing import strip_tags, strip_punctuation, strip_multiple_whitespaces, remove_stopwords, preprocess_string

from tqdm import tqdm_notebook as tqdm

## Read in data

# Read in articles
articles = pd.read_csv('NLT_data/merged_data.csv')

articles.shape

articles.head(2)

# Read in topics 
topics = pd.read_csv('df_dominant_topic.csv')

topics.shape

topics.head()

## Concatenate topics to original dataframe

articles_with_top = pd.concat([topics, articles], axis=1)

articles_with_top.head()

## Drop duplicate rows

articles_with_top_dropped = articles_with_top.drop_duplicates(subset='title')

articles_with_top_dropped.shape # 354 rows removed

# Save out as csv
articles_with_top_dropped.to_csv('articles_with_topics.csv', index=False)

## Train on Doc2Vec model
Taken from Clay's code in `gensim_doc2vec_recommender.ipynb` notebook.

articles_with_top_dropped.isnull().sum()

### Cleaning

df = articles_with_top_dropped.copy()

# filling some missing abstract values with the discussion section
df['abstract'] = df['abstract'].fillna(value = df['discussion'])
# Filling the rest with the title value
df['abstract'] = df['abstract'].fillna(value = df['title'])

# funtion for cleaning text
def preprocess(text):
    ls = []
    for i in tqdm(text):
        custom_filters = [strip_tags, strip_punctuation, strip_multiple_whitespaces, 
                          remove_stopwords]
        ls.append(preprocess_string(i.lower(), custom_filters))
    return ls

raw_data = df['abstract']

data = preprocess(raw_data)

### Model training

def create_tagged_document(list_of_list_of_words):
    for i, list_of_words in enumerate(list_of_list_of_words):
        yield TaggedDocument(list_of_words, [i])

data_tagged = list(create_tagged_document(data))

model = Doc2Vec(vector_size = 200, min_count = 3, epochs = 20)

model.build_vocab(data_tagged)

%%time
model.train(data_tagged, total_examples = model.corpus_count, epochs = model.epochs)

model.save('doc2vec_updated.model')

# Grab most similar articles of the input word and its most similar words 
def most_sim_docs(word, n_articles, topn=3):
    dfs = []
    sims = []
    
    # Input word with topn of its most similar keywords
    keywords = [word] + [w[0] for w in model.wv.most_similar_cosmul(word, topn=topn)]
    
    # Loop through each keyword
    for term in keywords:
        new_vec = model.infer_vector([term]) # random, different everytime 
        tag_list = model.docvecs.most_similar([new_vec])[0:n_articles]
        
        
        tags = [] 
        for num in tag_list:
            tags.append(num[0])
            sims.append(num[1])
        
        for t in tags:
            dfs.append(df.iloc[t, :])
        
    new_df = pd.DataFrame(dfs)
    new_df['similarity_percentage'] = sims
    
    return new_df

protein = most_sim_docs('protein', 2)

protein.iloc[1]['Keywords']

protein

## Assigning topic number to name

topic_list = [f"Topic {i}" for i in range(1, 26)]
topic_dict = {topic: float(i) for i, topic in enumerate(topic_list)}
topic_dict

# Search articles by topic name
articles_with_top_dropped.loc[df['Dominant_Topic'] == dic['Topic 4']]
